Modularize codebase, Update commentPopup function
