+++ src/index.html
@@ -10,7 +10,7 @@
 
 <body>
   <nav>
-    <ul class="d-flex partial0">
+    <ul class="d-flex j-space-around">
       <li><a href="#">BrightTop cinemas</a></li>
       <li><a href="#">Movies A</a></li>
       <li><a href="#">Movies B</a></li>
@@ -34,7 +34,7 @@
       </ul>
     </section>
   </main>
-  <footer class="d-flex partial0">
+  <footer class="d-flex j-space-around">
     <a href="#">Email</a>
     <a href="#">LinkedIn</a>
     <a href="#">Facebook</a>
